Officer recommendation part according status
