+++ src/services/api.ts
@@ -1,5 +1,3 @@
-// src/services/api.ts (or wherever you keep it)
-
 import axios, {
   AxiosError,
  type  AxiosResponse,
